requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.2.0
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
    #   django-cors-headers
async-timeout==4.0.3
    # via redis
billiard==4.2.0
    # via celery
boto3==1.34.152
    # via -r requirements.in
botocore==1.34.152
    # via
    #   boto3
    #   s3transfer
build==1.2.1
    # via pip-tools
celery==5.4.0
    # via -r requirements.in
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
distro==1.9.0
    # via openai
django==5.0.7
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-filter
    #   django-redis
    #   djangorestframework
    #   djangorestframework-simplejwt
django-admin-list-filter-dropdown==1.0.3
    # via -r requirements.in
django-cors-headers==4.4.0
    # via -r requirements.in
django-debug-toolbar==4.4.6
    # via -r requirements.in
django-filter==24.2
    # via -r requirements.in
django-nested-admin==4.1.0
    # via -r requirements.in
django-redis==5.4.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
dnspython==2.6.1
    # via
    #   -r requirements.in
    #   email-validator
email-validator==2.2.0
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
exceptiongroup==1.2.2
    # via anyio
gunicorn==22.0.0
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.8
    # via httpx
httpx==0.28.1
    # via openai
idna==3.7
    # via
    #   -r requirements.in
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jiter==0.9.0
    # via openai
jmespath==1.0.1
    # via
    #   -r requirements.in
    #   boto3
    #   botocore
kombu==5.3.7
    # via celery
numpy==2.1.1
    # via pandas
openai==1.75.0
    # via -r requirements.in
openpyxl==3.1.5
    # via -r requirements.in
packaging==24.1
    # via
    #   build
    #   gunicorn
    #   pytesseract
pandas==2.2.3
    # via -r requirements.in
pillow==10.4.0
    # via
    #   -r requirements.in
    #   pytesseract
pip-tools==7.4.1
    # via -r requirements.in
prompt-toolkit==3.0.47
    # via click-repl
psycopg2-binary==2.9.9
    # via -r requirements.in
pydantic==2.11.3
    # via openai
pydantic-core==2.33.1
    # via pydantic
pyjwt==2.8.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pytesseract==0.3.13
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   botocore
    #   celery
    #   pandas
python-dotenv==1.0.1
    # via -r requirements.in
python-monkey-business==1.1.0
    # via django-nested-admin
pytz==2024.2
    # via pandas
pyyaml==6.0.1
    # via -r requirements.in
redis==5.0.7
    # via
    #   -r requirements.in
    #   django-redis
requests==2.32.3
    # via -r requirements.in
s3transfer==0.10.2
    # via boto3
six==1.16.0
    # via
    #   -r requirements.in
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlparse==0.5.0
    # via
    #   -r requirements.in
    #   django
    #   django-debug-toolbar
tomli==2.0.2
    # via
    #   build
    #   pip-tools
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   anyio
    #   asgiref
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
tzdata==2024.1
    # via
    #   celery
    #   pandas
unidecode==1.3.8
    # via -r requirements.in
urllib3==2.2.2
    # via
    #   -r requirements.in
    #   botocore
    #   requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
wheel==0.43.0
    # via pip-tools
whitenoise==6.7.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
